@@ -1,7 +1,6 @@
 package main
 
 import (
-	"io/ioutil"
 	"log"
 	"os"
 	"path/filepath"
@@ -16,14 +15,14 @@ func main() {
 
 	log.Println(cwd)
 	p := filepath.Join("internal", "generatable", "_generated")
-	files, err := ioutil.ReadDir(p)
+	files, err := os.ReadDir(p)
 	if err != nil {
 		log.Fatal(err)
 	}
 	for _, file := range files {
 		log.Println(file.Name())
 		fileP := filepath.Join(p, file.Name())
-		b, err := ioutil.ReadFile(fileP)
+		b, err := os.ReadFile(fileP)
 		if err != nil {
 			log.Fatal(err)
 		}
